+++ test/index.test.ts
@@ -9,7 +9,6 @@ import contentSecurityPolicy from "../middlewares/content-security-policy";
 import crossOriginEmbedderPolicy from "../middlewares/cross-origin-embedder-policy";
 import crossOriginOpenerPolicy from "../middlewares/cross-origin-opener-policy";
 import crossOriginResourcePolicy from "../middlewares/cross-origin-resource-policy";
-import expectCt from "../middlewares/expect-ct";
 import referrerPolicy from "../middlewares/referrer-policy";
 import originAgentCluster from "../middlewares/origin-agent-cluster";
 import strictTransportSecurity from "../middlewares/strict-transport-security";
@@ -24,7 +23,7 @@ import xXssProtection from "../middlewares/x-xss-protection";
 describe("helmet", () => {
   const topLevel = helmet.default;
 
-  it("includes all middleware, except Expect-CT, with their default options", async () => {
+  it("includes all middleware with their default options", async () => {
     // NOTE: This test relies on the CSP object being ordered a certain way,
     // which could change (and be non-breaking). If that becomes a problem,
     // we should update this test to be more robust.
@@ -34,8 +33,6 @@ describe("helmet", () => {
       "cross-origin-embedder-policy": "require-corp",
       "cross-origin-opener-policy": "same-origin",
       "cross-origin-resource-policy": "same-origin",
-      // In Helmet 7, we can remove this Expect-CT assertion.
-      "expect-ct": null,
       "origin-agent-cluster": "?1",
       "referrer-policy": "no-referrer",
       "strict-transport-security": "max-age=15552000; includeSubDomains",
@@ -62,22 +59,11 @@ describe("helmet", () => {
     });
   });
 
-  // In Helmet 7, this test should be removed.
-  it("allows Expect-CT to be enabled", async () => {
-    await check(topLevel({ expectCt: true }), {
-      "expect-ct": "max-age=0",
-    });
-    await check(topLevel({ expectCt: { maxAge: 123 } }), {
-      "expect-ct": "max-age=123",
-    });
-  });
-
   it("works with all default middlewares disabled", async () => {
     await check(
       topLevel({
         contentSecurityPolicy: false,
         crossOriginEmbedderPolicy: false,
-        expectCt: false,
         originAgentCluster: false,
         referrerPolicy: false,
         strictTransportSecurity: false,
@@ -295,9 +281,6 @@ describe("helmet", () => {
     expect(helmet.crossOriginResourcePolicy.name).toBe(
       "crossOriginResourcePolicy"
     );
-
-    expect(helmet.expectCt.name).toBe(expectCt.name);
-    expect(helmet.expectCt.name).toBe("expectCt");
 
     expect(helmet.originAgentCluster.name).toBe(originAgentCluster.name);
     expect(helmet.originAgentCluster.name).toBe("originAgentCluster");